TeX documentation and new variable
Wrote the documentation of the completed geometry and some notes on the functions that the GUI uses in TEX. Introduced a new variable, so that arrowhead sizes can be changed as requested. Added the Icon for the GUI (code broke when I run it on other machines, I then needed to # the line out). Let me know if you need/want the TEX file also, to update it as we go along.