changes to polar plots
Still nowhere near the desired outcome. But now it puts conversions of polar fields in the x and y component boxes, so that the components under the PLOT button can still be used. The polar field submits upon choosing the vector type in its separate window. The problem that remains is the limited use of arctan in numpy, and the problem of replacing variables  (names had to be altered to be somewhat unintuitive).