House Keeping 2
This is significant at this point. Due to the quick inflation of the code and files, we need at this point dashoard to run/access them all. I made a initial one -dashboard.ipynb-. All files can be ran from the dashboard and with option to order the codes from the dashboard. The files names changes are important for the load from the dashboard and they make sense I guess