Both ways of visualising 2-forms, as discussed
I have attempted to implement both methods of visualising 2-forms that we talked about. Not sure if these are great ways of representing them, will try to think of some other ones also. Let me know if putting them in a separate folder creates any issues, apologies if it does.