GUI now accepts fields with no x and y dependence properly
Sam has found that the GUI did not accept constant fields. There was an issue of checking if components are zero, but not checking if they are integers. When both components were integers, the code crashed because some other parameters were not longer arrays are were being sliced. I fixed it by checking is x and/or y are present in any of the input fields.