Included a dropdown menu
I added a simple dropdown menu, it is surely not the most efficient way of doing this. But updating it from here should be very simple and it works completely fine. I did not add all the fields the VFAII has because they use a command 'box' that I am not sure how they define.