removed 'a' and made the window scale to full-screen on all machines
As we mentioned, I removed a. I also made the size of the Tkinter window extract width and height from screen information. The window should be full-screen sized when it opens now. The probleem of frame paddings will still remain as we add functionsm but It can be designed nearer the end.