changed to np.arctan2, fixed the issue with polar stacks not plotting
Tried using np.arctan2, it works better, but for some fields, range is still an issue, better though. Also fixed the issue of stacks not plotting in polar. To scale the polar field without having to convert to a cartesian grid, I added that scaling to the polar plot window.