2 forms on R3 attempt 1
I have attempted to implement visualisations of 2 forms on R3. So far, it does not work correctly. It works when viewing along the z axis, for any value along it. But it does not work (yet) for viewing along the x and y. I'm not sure why, will spend more time on it and improve it.