Very important fixes to int deriv
The interior derivative was completely wrong. Now it plots the resulting 1 form in terms of stacks. As the end goal, once this is ready, is to merge it with the main code (plotting_stack_GUI). I have not implemented customisations, these will happenautomatically in main code through global variables. I will merge the 2 once this code is satisfactory.