slider change and variables name
changed the z_index to a more intuitive, h_index. z_index was confusing as, when viewing along the x or y axis, it was still called z_index, but not corresponded to the height in x or in y no longer z. More importantly, I changed the slider to update when the mouse is released. This gets rid of the clumsy button and overcomes the issue of updating with every unit change, making the code slow. What I will now try is to display actual heights, not indexes along the viewing axis. More updates very shortly.