Slider indexes and heights
Slider used to display only the index along the viewing axis. I tried making it display the value instead in many (some very convoluted)  methods. None worked. The only way I have found is this:
This code shows the index from the slider above (labelled)  and the value below.
The problem was:
--- I could not get the slider to go up with axis values, I have found methods of doing this, but they are very messy, slow, and unclear.
--- I attempted disabling the index showing by the slider and instead just putting down the value along the axis manually, using labels. This works, but only updates one the slider is let go of and a choice is selected. The numbers don't follow the mouse.
--- I also tried to get the slider to select values equal those in arrays (based on L and pt_den), this required getting h_index from h_value. Still it did not work, because the slider resolution does not update properly, it rounds, so no h_value could ever be found.
--- with this code, the indexes follow the mouse and once its selected, a value shows underneath.

IT IS NOT WHAT I WANTED, BUT IT WORKS FOR NOW, I WILL KEEP WORKING ON THIS TO UPDATE MORE CLEANLY. If anyone has any ideas of how to do this, feel free.