R3 possibly fixed (?)
I think this may have fixed R3. I'm finding it hard to confirm it (hard to visualise it before plotting). I checked it for constant and linear fields, think it works. Also confirmed that it agrees with the 2 form on R2 plot.

Again, please let me know if you spot anything odd about it.