Fix for constant fields, still not working
Fixed it for constant fields, they produce the same fields now.
PROBLEM:
When cancellations occur after differentiation, the 1 forms are unchanged and therefore despite all magnitude being correct, the superpositions of stacks, are not. Need a way of detecting if results will cancel, therefore testing if the two derivatives contain any common parts.
When common parts exist, could exclude them from the stacks expression.
I am working on implementing that now. Not as easy as I thought it would be.

PS The code loads with an example that breaks due to cancellations.