First attempt at calculus with 2 forms and 1 forms GUI
This would be the first attempt at 1 forms and 2 forms calculus.
Much still does not work:

-- Hodge is not yet implemented
-- Wedging does not work fully
-- 2 forms from given expression and even from supplied 1 forms via the exterior derivative do not work, they are not unique solutions, and the only other way I have found includes taking care of cancellations between two expressions before they are merged.

This file will (hopefully) later be included into the main GUI. As that will make the code very long and clumsy, this will only be attempted once all this calculus is known to work on its own.

I will still be working on this. Let me know if you find any improvements and/or other, not yet mentioned issues.