No major changes
Can't find a way of disabling the selected toolbar options when 'tools' is deselected.
Before changing from 'tools '  to anything else, the options need to be disabled, otherwise it breaks. The user can do it, would be great if it disabled whatever is selected automatically, to avoid issues. Won't work though. Even destroying the toolbar and making a new one doesn't disable made selections.