More on dealing with singularities
Happy with isnan method and I think it's mathematically more sound than the box thing. Just ameded  Sam's section that deal with them since if any part of the field is not defined in a region (u or v) the whole field shouldn't be defined there. Previousely, if u is causing trouble and v is fine in certain region, then the field will be plotted in the region that's banned for u and vise versa. Example: ln(x)dx+ln(y)dy should be plotted only in the fully +ve quadrant, it was graphed over three quadrant  due to the mentioned issue.  This  will hopefully close this problem.