Hopefully, complete fix to toolbar
Toolbar should now work all as expected.
It should be available to use in the 'tools' option. When anything is selected, it should go back to home, unselect any option that was chosen (pan or zoom) and make the buttons disappear. When selecting back to 'tools' it should make them appear again.
It is still showing an error about some 'message' attribute, but it seems to not be important.
Needed to delete initial plot of chosen option (deriv, curl, div etc) because it sometimes broke. When code was run for the first time after refresh, x_pix and y_pix would not be defined until the figure is pressed, but if before any figure presses, any option was clicked, deriv was called. and derriv needed x_pix from a click that did not happen yet.

Should hopefully work now. Let me know if there are still any issues with it.