Some customisations
These are not perfect, and needn't be.
This code I plan to merge into the main GUi in which this step will only include putting one extra box in its customisation box.
The other code has nice functions to replot whaever is on screen with new parameters that can be used by this section when the codes are merged. For now I implement it like this here just to test it.