Another Singularities update, more to come
This one will recognise if the singularities that lie on our grid are a point or on a line and plot them as red dots.
If these are regions, it will shade them.
Try:
ln(y)dx + ln(x)dy
or
Gravitational field of a point charge from dropdown
or
1/(1-x) dx + 1dy

NOTE
Nothing works when singularities do not lie on gird points, say, is there is a singularity at 0.3 or pi or 0.1234 etc.
Im still thinking about how I could implement that.