singularity regions hyperbolics
Found some cases for which the region finder broke. Fixed them now.
Included hyperbolics and inverses of trig.
Not sure if want to keep it as it is, for now I got around the issue of function names being included in other functions by using capitals. Note, inverse hyperbolics still will not work. A better method is required there.