More toolbar
Changes:
This version removes the pan and zoom tools when first selecting an option other than tools, but these tools don't return later, as toolitems list in NavigationToolbar2Tk has been altered. (can be seen in console if you enter toolbar.toolitems) If there is a way to restore these items then we should be sorted.