Singularities and polar
Changed the way we recognise singularities to be just as python labels them instead of basing them on regions. Not ideas, python has its problems with confusing nan and infinity at times. But better than the regions method (more consistent).

Changed the polar grid to scale with L. pt_den is not used when polar plot shows, so it now colours red when that is the case.

Added new pre-defined fields: black hole filed.

A good example of when python still breaks with nan vs inf is the black hole ect field at minimum radius = 0, number of points along r=11, number of points along theta=30, size=5.