More on LI
Line integral getting better.
Ideally, would also want it (as well as deriv. curl and div) to work in polar too.
But now its displays it on a label, updates in real time, resets when other options are chosen and has a reset button.
Ideally, would want reset and other Radiobutton selection to just delete the lines, not just replot. But for now it works.