Some singularities
Includes a check for singularities on a fine grid of specified number of points. Up to 200  points is still quite useful and rather fast for most fields.
Includes an expression input on which the user knows a singularity to exist. As lines from equations and as point coordinates.

I have also been working on the step function way of dealing with singularities but am struggling to make it useful. Will be continuing with that for today.