Singularities
For the case of plotting singular points, it now warns the user if the input point is actually singular or not

Problem with the lines is that, there are many values, requiring all to be singular is not consistent, due to bugs and python rounding. For now, I imposed that if any point along that line is a singularity, I don't display the warning

Let me know if you have any improvement ideas.