Some documentation
Wrote some documentation.
Some more about R3, section about singularities and most importantly, a section of how we now plot 2 forms. The sections on our previous methods for this are kept, but I modified them to past tense and with clear highlights of issues.

I struggled to clearly, in a single structured argument present why exactly this method is more accurate as a 2 form representation. I write down how I think of it and why it seems better, but reading it, the description is not the best. If any of you can tweak it to be clearer, would be much appreciated. Thanks!