Documentation update
A new format to mirror the potential work publication. I mainly worked on chapter 4, so main changes and additions are there, and couple of things here and there. I will try to work on the rest before our next meeting. This document will be the main source for the paper, so the more details we add the better. Cheers