'Fixed' exp, and other
When I say 'fixed', I removed it. The input, everywhere is e**() or e^(), all else breaks. This is because SymPy simplify removes brackets when there is only one argument in the exponential, which then gets formatted and the brackets are not there, so NumPy breaks. I updated the documentation to include that.

Also:
--- fract now changes also when selecting 1-form plot and 2-form plot in Calculus window.
--- I changed all names in code from '2 form' to '2-form' and from '1 form' to '1-form'.  Done the same in documentation too.
--- fixed an error with not disappearing buttons in LI tab.