Initial 1-form to VF distinctions
Includes the distinction between VF and 1-form in main tab.
When stacks are selected, all labels name 1-forms, div curl and deriv are disabled. When arrows are selected, everything returns to as in VF and radio-buttons activate again. Also fixed an issue with mouse still doing what was selected (div, curl etc) all tabs. Fixed an issue with plot buttons not returning to normal after selecting calculus or R3, and going back to main or LI.

Fixed a new found issue too, LI now has its own radio-buttons for stacks or arrows. These link to same functions therefore create no new inconsistencies with the first tab (as far as I could find). When both is selected, as that is useless for LIs, it is automatically changed to stacks.

I will be getting on with other VF / 1-form distinctions and a zooming tool for 2 forms now.