Set target
Implemented the Set Target for inset axis in the main tab and in calculus, as in VFAII. The pixel coordinates I need for this are not exact (did not find a way of extracting these yet), but are VERY close.  Can now see that zeros do plot in the zooming of 2 forms. Unfortunately, not at all positions (rounding errors).