dropdowns
The 1-form/VF dropdown menu in the bottom frame now changes to stacks automatically (with all the limitations that come with that change) when the Electric field, H-field (changed name), Flamms parabloid or BH are chosen.

I include a dropdown menu for 2-forms on R2, in the Calculus tab. For now, not included many, just tested if this works.
Also I merged the plot frame with the bottom frame to give us some extra space for the option frames on the right. Technically, these are just separate notebooks, not a separate frame now, so that correct bits disappear when going into R3. We can work on the cosmetics in the last two days, as to the R3 frame especially.

@MostaphaG please let me know if it vertically fits on your screen now.