Integrals of 2-forms
The integration should mostly work now.
It is a bit slow because it has to evaluate fairly large grids. It does not work with singularities too well. (kind of expected). Haven't tested too much but work ok for all I tested. Also, because of the speed limitations, it is not the most exact.
What is quite nice is that it shades the area it integrated over. I used it for testing but kept it after, I like it. I think its slowing the code quite a bit though, let me know if we want to sacrifice it for speed.