Optimisation and shade option
Somewhat optimised the Area integral. Also, the shading is very nice but it is the prime cause of this code being slow. I added a toggle button with which a user can choose whether they wish to shade and wait a little longer and not shade and just get the answer fast.