Updating Dynamics
Changes:
Added buttons to stop, pause, clear the current points in the animation.
Added sliders so user can determine the max time duration of the animation and the number of frames used in the animation.
Next would like to have shapes so we can see how the shape/size changes in the vector fields.