Some minor cosmetics
Changed order of frames and their names.  Did nothing to the documentation, just run it. but for some reason it still shows a change.  Changed the layout of the bottom and tried my best to reduce the white space around the figure. Still not figured out how to make Tkinter the same size for all of us, could not find much online yet. @MostaphaG  let me know if it still fits on your screen.