Attempts at circle
IT DOES NOT WORK.
Tried really hard to get this working but the matplotlib pacthes method 'conatins_point' seems really bad when it comes to circles. I can do squares with this code. It does not check if the point is within circle, it just checks if its within a square bound by it... The red line shows the circle the patch defines, the grey shows where it evaluated. I'm surely missing something, but I've spend very long on this and still cant get it to behave.