More shapes in dynamics
Implemented some shapes in dynamics
Include a square and a human. Not yet included a circle. The way to use it is to set the size, click on the graph (this will plot it), then press animate.
Would like to remove the red dots when human is drawn, makes it clearer. Could not find where these are set though..
@skirwinjones, if you remember, feel free to implement that.