tmax and spped attempt
Can't quite get this right. The issue is the function we use to animate. Normally we just set total number of frames and interval for how long each frame displays, from that we get the time by t = frames*interval/1000 (because interval is in ms). Can rearange that, then interval gives the speed, and by that the the time the user gives, we get how many frames we need. The problem here is the time array, it seems to mess with it all, now t_max depends on interval somehow and is always longer than what the user wanted. Not sure how to go about that issue. Now the speed does change the speed on some level and time implied duration, but its not linear.