Showcol and Analytical Div&Curl
Changes:
Showcol now changes the circle automatically without requiring additional click.
Div and Curl are computed analytically at the the click location and displayed in the inset window. What do you guys think about this? Perhaps add a button to show/hide this text? Or move into the corner of the inset plot?