Concept
Mostly just corrected some minor bits, like the toggle for the torus disappearing after anything but the polygon is selected. Thought about hover tooltips. Tkinter has an outdated module that deals with this, so far I did not find an up to date one. Will keep looking and will note that on the to-do-list. Here I used an idea we could use for the future.

Each button that needs an explanation can be hovered and that will display instructions in a separate frame (@MoustafaG, I know that so far you will not be able to see it on your screen, but I've just put it in the code for the logarithmic scaling toggle, for now its only there to keep the idea in mind, if we sort out the Tkinter on MAC problem though, we can easily use this). However, this would get the code quite long, with all these functions, each putting on a label for its corresponding button. We could simplify it a great deal (to actually be quite neat in the end).

We can end up with just one simple function. The idea being, we can store these instructions as text in a file, each with its own index. When a button is hovered-over, it goes into a response function with 2 entries, index for the instructions, and 0 or 1 (Hovered into button, or out of) The function will just access the file and display the right label, for each button, quickly, these will disappear when uncovered.

Let me know what you think of it. Or we can discuss that tomorrow on the meeting