Adjustment
Slight adjustment to previous push. All remains the same but the frame is now a notebook in the bottom frame. All issues and future aspects as described in previous push and in to-do-list. Problem before was MAC users not seeing the labels, now, when labels display, it will push all options off the screen. Also, some might already be off the screen on the right...