dynamics 1 bug fix
Fixed the bug with it breaking every few runs, noticed that it included some parameters never being deleted and then being accessed before changing them to new values. The common error of 'NameError: xy0 is not defined' should not be a problem anymore.

Noticed a big issue. The fucntion we use for diff eqn solving (RK4calc) is very inefficient for out needs, it calculates the whole trajectories of all particles and only then animates. For long time scales (try setting interval to 0.1 and time to 10s) this breaks becuase it has too much to calculate and it crashes.

Fixes:
---We either need a better method for the solving (one that does it interval by interval, possibly write our own)
--- OR we need to thread the calculation, so it can happened without crashing the code.