Hover instructions
Hover instructions generalised, to add new ones now, just add a line with it to the text file and bind the button to the function with its index. I've done it for the log scaling toggle and the Autoscale toggle. The text is sliced automatically to new lines, not to exceed the width of the right frame.
I also removed the outer frame boundaries and changed 'Torus' to 'PAC-MAN grid'