Annihilate in first tab
Included an option to annihilate the 1-form and the vector field when both is selected. This plots the contours of the resulting scalar function. The equation for it is not showing anywhere. There is no reason to have a 0-form box in the VF tab. A label for that could easily be done, didn't see the point in that though. Let me know if you think it is something vital. I do describe that function in the new instructions frame, and I do mention that the two will annihilate even if they are not equal and point the user towards the interior derivative.