Some changes from the To-do-list
Increased the font size of tabs to 11.
Implemented the VF and 1-form language changes to all tabs now, not just the first one ('VF' tab)
Made the right frame expand over the side of the bottom frame, now when the button for log scaling is hovered, the instructions appearing should not displace it away from the mouse. In fact, it should not be displaced at all now.
Got rig of the label appearing above the bottom frames when changing to VF options
Changed x_m and y_m to being as zero and not as None to stop some errors appearing in first time use (rare)
analytic toggle when selected on tools was showing an empty inset and and error, fixed that too.