Started on the library
Have an initial, usable version of stack_plot in library form. It takes in just some basic components it needs and plots a 1 -form over its defined range. If the user wants to change more details, eg. showing arrows, autoscaling, arrow orientation etc. They can do that using callings. It has an example use in it at the end, for now. Later we would probably want such examples in a separate file that imports this library.

Main problem. In order for it to use a figure that the user has pre-defined, it will need to act as a method, on a matplotlib object. To do that (I think) we would need to add it to the figure matplotlib class, which I don't think we can do. For now, the function creates the figure for the user and defines the range using the maxima in arrays. The user can change its properties by axis methods. An example of that is implemented in the example.