Added a way of defining a 0-form object
Added 0-form object. It plots contours. Mostly works in the same way as previous. Has an option to clear axis before replotting or plot on top. Might add that to the other ones too.
Noted some issues and possible solutions for the future:

Most of our calculations use strings and there is no obvious way to get around that. eg denser grids on contour plots that from original form, all of out ext. alg. is based entirely on strings etc.

To get around that on the 0-form, I added an option for the user to supply the equation. Once that is done, some new options will start working. Otherwise, they will break because the string has not been given.

The string is given as a self not as an input to method as it will get reused a lot, it will be easier this way.

I will later implement this to the 1-form and 2-form. Once that is done, we can start implementing exterior algebra that will use these. The idea I have with those is to make them methods on the form object, depending on what object they get called on, they will get the result and call the class to set up that new form.

Eg. function that defines a 1-form will have a method for an exterior derivative. That function will be from our main code. Once that is calculated, the result is a 2-form. So at the end, it will supply its result to our function that sets up a 2-form object, this object will be returned by that method. From there, the user can use all 2-form methods on it, including plotting.

I will start implementing that tonight/tomorrow night. Let me know if you have any issues with this.