Hodge of 1-form and fixes
Included a Hodge of a 1-form. Numerically and analytically, user can choose. I also fixed an issue with constant 1-forms from exterior derivatives.  All comments needed to understand the Hodge are there. Also implemented some more examples for everything I added.

Crucial: I added a way for the user to supply a figure into the form object if they so wish (instead of it creating a new one for itself). Useful when the user wants to update things one the same figure between different objects. Seen best on the last example.