Subplots and Big Error
Added ways of dealing with subplots. Got somewhat confusing though so this section will need extensive documentation. But it should allow the user to do a lot. Add them onto the figure, supply ones to start with. These should pass between objects via the exterior algebra methods so they shouldn't have to be redefined. Etc.

Found a major mistake in the GUI. Tried a 2-form - (-cos(x) -1) and it broke badly. Found out why that was, fixed it. Should be fine now.

Will hopefully move onto a Wedge product next.