More methods
Implemented a Hodge of a 2 form and an analytical wedge of 1 forms. I tested the 1-forms wedge, seems to work ok (not extensive). Would be good to have a numerical one too though.

I have not had the time to test the 2-form Hodge and to write examples, I will try to get that finished tonight. Should work ok though (I think).
I will also try to get the 0-form Hodge done tonight. If I manage for time.

If anything from this push is not working, don't worry, I will fix it tonight.