Fixed issues from earlier and Hodge of 0-form
Fixed issues I made earlier, when I run out of time to finish my work.

I added a numerical and analytical Hodge of a 0-form, in the same way as Hodge of a 2-form. Included an example for that too.

I deleted the vector field to 1-form and vice-versa via the metric for now. Need to think about that some more before I implement.  I think its very much doable though.