Interior derivative of a 1-form, numerical and analytical
Implemented the numerical and analytical interior derivative of a 1-form with very flexible input for the VF.  Returns a 0-form object. Also, made sure that all functions we coded before, deal with constant components (quite some did not).
Included examples of all options I give to the user, apart form passing on figures and subplots, but these work in the same way as before.

I will get onto the 2-form interior derivative soon, hoping to get that done by Tuesday.  Should not be too bad now that I have this one working, I know how to deal with them.