Involving the metric
I included some code to let the user change a VF object to a 1-form object. It does that through the metric, which can either be supplied with equations, or as numerical grids. It can work analytically, numerically, or both.
I'm struggling to test if it is giving expected answers as I don't have great intuition as to what they should look like. But I'm pretty sure that, in terms of equations and comparing to 1-forms set up with the result equations to begin with, all is fine.
@MostaphaG  It could definitely do with some more testing.

I am yet to do the same the other way around, transforming a 1-form to a VF. Now that I got a hang of it this way, the other way around should be ok. I will try to find some time to do that very soon.