Tried some testing with singularities
Tried removing nan and changing infs to manageable values, not only when plotting, but also before recalculating the NumPy gradient. I thought that would help, as it was, we were doing NumPy gradient on it with nans and infs in there. It does not seem to help. Will need to think about this some more. Kept the changes but commented them out (in formpy).