1-form to VF via the metric + extra
Included a method to change the 1-form into a VF, via the metric. Same method as for VF to 1-form. Tested it with general entries, works ok (no errors). Then tested it with the interior derivative (re-did the maths to check it backwards) and it still produces expected results.

Also, played around with the numerical exterior derivative, did some reading into numpy.gradient, couldn't find any issues for us from that. Tried removing singularities etc. I think I might have a possible further idea:

I tested the num ext_d with other, simpler fields which have singularities in them, and they work fine (mostly, to some expected differences, like not drawing singularities on when I explicitly remove them). I'm starting to think our issue with the B field might not be the singularities or intrinsic to numpy.gradient, but rather, something to do with the specific form of the function we give it. Just a guess so far though.