Major fixed in zooming
zooming with zoom =1 did not work. This is because, actually, never did the inset correspond to the correct size in the figure. Even if we don't end up merging this branch in, must include this change to the original FormPy.

@MostaphaG  @skirwinjones

BTW, sorry for the push spam. Been quite productive today, so there has been a few things to report.