Testing and error handling
Changes:
Trying out some examples with the plotting and zooming. Happy with the results. Added some error handling for vector field so that user cannot enter invalid values for zoom and insize (new name for d_length - not sure how I hadn't noticed how dodgy that sounds up til now!!)
Suggestions:
Perhaps we create a new figure if an axis isn't specified, so that there is a kind of 'plug and play' functionality for less experienced users?
Can we add a feature that adds a default colourbar for 0-forms so the user can see more clearly the numerical values of the contours?