Simple changes
Changes axis names for Lorentz example.
Also, made default of zooming to do so on an inset. I know its not ideal with way, because it requires axis anyway, but the user will more often use it with inset=True anyway and them give an axis. One thing less to write.