Co-contra example and camp
Added cmaps to be handled by 0-form contours. No method to change it, but can do it directly by accessing the variable like I do in the figure code.

Creates the initial figure proposal for the co/contra variant example, using the 2D BH metric. Not put in the paper just yet.