Example code
Wrote some example code for zooming on a 1-form. To put it down before the VF on the paper with code and then have the Vector field, with no code example.

Also, changed the figures for VF to be horizontal as they will not have code anymore, and 1-forms to be vertical, they will not have code next to them.