Log for vector fields
Changes:
Added a method for log scaling VF
Scales the magnitudes of the vectors themselves, so autoscale etc still works in the same way. Added an example to test in sam_examples and tried it with the Lorentz figure.

Problem: If the VF is constant then the resultant field after log scaling is zero. :( So should probably use a different method.