Classes not functions
Classes in python are importable, no need for overlay functions. Importantly, this will allow us to check against the type of each class, by eg. type(form_obj, form_2). That example would retrun True if form_obj was a 2-form instance.
More importantly isinstance(form_obj,  form_2) would work too.

IMPORTANT @MostaphaG @skirwinjones, In the process I discovered that set_density no longer works for a 2-form (and 2-form only).  I checked, it is not because of the naming of the class and the self instance both being the same (form_2). Will need to investigate to fix that. I do not think the removing of functions did that, but I can be mistaken.