More analytical wedges on 1-form plus more
Right, bear with me here @MostaphaG @skirwinjones
Trying to get more wedges in (here: 1-form/\0-form and 1-form/\2-form). I only implemented it for for the analytical wedge acting on a 1-form. Examples In test file. Its quite time consuming to change it to work like this. But I will endeavour to do it for all methods.

Once I am done I will explain how it works in the paper. The general idea is, for 1-form/\1-form, user can give us a 1-form instance, nothing, or a tuple fo2 equations. For 1-form/\0-form, user gives us a 0-form instance, or a single string for 0-form equation and tells us the degree of the form they wedge with, so degree=0 here. To do 1-form/\2-form, user would give a 2-form instance, or a single equation with degree=2, but this one will only show a message that the result is a 3-form, which on R2 is always =0.

Takes some time thinking this stuff, in terms of the code....

I also implement an exterior derivative of a 2-form, just showing the same message.

Oh, and a typo in interior_d I noticed that would break if code got to it (self. missing in 2 places and one xg was supposed to be yg).