Readme update
We are couple of setps from publishing the library. Just few things that you need to check most of the file for the library package are self explanatory. Main one is setup.py. I thought our development status is 5 - Production/Stable, check here https://martin-thoma.com/software-development-stages/ for details and let me know if you think we are in different state. For testing and update, please keep updating the original file in Library folder. All changes will be added before we publish.