setip.py edit
Just resending this so it doesn't get lost. @MACIEK1JAREMA, @skirwinjones We are couple of setps from publishing the library. Just check the files in the FormPy_Library,  most of them are self explanatory. Main one is setup.py. I thought our development status is 5 - Production/Stable, check here https://martin-thoma.com/software-development-stages/ for details on this parameter, and let me know if you think we are in different status. For testing and update, please keep updating the original file in Library folder. All new changes will be added before we publish.