Fix for irregular 0F
Important (good spot Maciej): The problem is that we have used i and j the wrong way round every time we are looping over array elements. The first argument denotes the row and therefore should correspond to y in the plot. I have fixed for 0F but I am going out for dinner so will have to resume tomorrow. I fear that it may be a more complicated issue where the plotting of stacks is involved.