More rigorous testing
tested everything with singularities, not square and not equal size grids.
@MostaphaG @skirwinjones  This stuff is starting to work amazingly well now. Love it.

The ONLY thing I have found breaking is insets. They get placed not precisely where 'target' is. It's always a bit off.
I tried looking into it, but I can't find why it does that. I didn't code this initially though, so I don't fully understand how it works. On a plus side, it handles odd grids fine, that drift does not get larger.

Can @skirwinjones please look over that soon? It does that for all objects (VF, 1-form and 2-form)

As far as my testing went, this is the only remaining problem now.
(And perhaps the 0-form log scaling not being the most effective, but as far as I could find, its not just our issue, matplotlib seems odd with contour log scaling too)

Apart from that, I think this baby is ready to publish.